Improved entity mapping, adding recipes still does not work

+++ CRUDBackend/src/main/java/com/example/crudbackend/Repositories/IRecipeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface IRecipeRepository extends JpaRepository<Recipe, Long> {
     Iterable<Recipe> findAllByNameContaining(String name);
-    Recipe findByRecipeid(int id);
+    Recipe findByRecipeid(long id);
 }
